refactor(cli): extract skills/agents/plugins/sandbox/models/usage handlers out of daemon.ts
Phase 1a of splitting runDaemon()'s 2148-line closure into per-concern
modules (daemon-handlers/). These six sections only needed `socket` +
an isShuttingDown() getter, so extraction was mechanical — no shared
context object needed yet. daemon.ts: 2442 -> 2055 lines.
