+++ packages/cli/src/runner/daemon.ts
@@ -41,35 +41,21 @@ import { defaultStatePath, acquireStateAndIdentity, releaseStateLock } from "./r
 import { normalizeLoopbackHost } from "../relay-url.js";
 import { startUsageRefreshLoop, stopUsageRefreshLoop } from "./runner-usage-cache.js";
 import { startOllamaModelsRefreshLoop, stopOllamaModelsRefreshLoop } from "./runner-ollama-models-cache.js";
-import { getWorkspaceRoots, isCwdAllowed } from "./workspace.js";
+import { getWorkspaceRoots } from "./workspace.js";
 import { type RunnerSession, spawnSession, killSessionProcessGroup } from "./session-spawner.js";
 import { pruneSessionCloseMetadata, type SessionCloseMetadata } from "./session-close-metadata.js";
 
-import {
-    scanGlobalSkills,
-    readSkillContent,
-    writeSkill,
-    deleteSkill,
-} from "../skills.js";
+import { scanGlobalSkills } from "../skills.js";
+import { scanGlobalAgents, readAgentContent } from "../agents.js";
+import { scanAllPluginInfo } from "../plugins.js";
+import { initUsage, triggerScan, closeUsage } from "../usage/index.js";
 
-import {
-    scanGlobalAgents,
-    readAgentContent,
-    writeAgent,
-    deleteAgent,
-} from "../agents.js";
-
-import {
-    scanAllPluginInfo,
-} from "../plugins.js";
-
-import {
-    initUsage,
-    triggerScan,
-    getData as getUsageData,
-    closeUsage,
-} from "../usage/index.js";
-import type { UsageRange } from "../usage/types.js";
+import { registerSkillsHandlers } from "./daemon-handlers/skills.js";
+import { registerAgentsHandlers } from "./daemon-handlers/agents.js";
+import { registerPluginsHandlers } from "./daemon-handlers/plugins.js";
+import { registerSandboxHandlers } from "./daemon-handlers/sandbox.js";
+import { registerModelsHandlers } from "./daemon-handlers/models.js";
+import { registerUsageHandlers } from "./daemon-handlers/usage.js";
 
 // Re-export migration from shared module — used on daemon startup
 import { migrateAgentDir } from "../migrations.js";
@@ -1506,386 +1492,13 @@ export async function runDaemon(_args: string[] = []): Promise<number> {
             (socket as any).emit("pong", { now: Date.now() });
         });
 
-        // ── Skills management ─────────────────────────────────────────────
-
-        socket.on("list_skills", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const skills = scanGlobalSkills();
-            socket.emit("skills_list", { skills, requestId });
-        });
-
-        socket.on("create_skill", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const skillName = (data.name ?? "").trim();
-            const skillContent = data.content ?? "";
-
-            if (!skillName) {
-                socket.emit("skill_result", { requestId, ok: false, message: "Missing skill name" });
-                return;
-            }
-
-            if (!/^[a-z0-9][a-z0-9-]*[a-z0-9]$/.test(skillName) && !/^[a-z0-9]$/.test(skillName)) {
-                socket.emit("skill_result", {
-                    requestId,
-                    ok: false,
-                    message: "Invalid skill name: must be lowercase letters, numbers, and hyphens only",
-                });
-                return;
-            }
-
-            try {
-                await writeSkill(skillName, skillContent);
-                const skills = scanGlobalSkills();
-                socket.emit("skill_result", { requestId, ok: true, skills });
-            } catch (err) {
-                socket.emit("skill_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("update_skill", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const skillName = (data.name ?? "").trim();
-            const skillContent = data.content ?? "";
-
-            if (!skillName) {
-                socket.emit("skill_result", { requestId, ok: false, message: "Missing skill name" });
-                return;
-            }
-
-            if (!/^[a-z0-9][a-z0-9-]*[a-z0-9]$/.test(skillName) && !/^[a-z0-9]$/.test(skillName)) {
-                socket.emit("skill_result", {
-                    requestId,
-                    ok: false,
-                    message: "Invalid skill name: must be lowercase letters, numbers, and hyphens only",
-                });
-                return;
-            }
-
-            try {
-                await writeSkill(skillName, skillContent);
-                const skills = scanGlobalSkills();
-                socket.emit("skill_result", { requestId, ok: true, skills });
-            } catch (err) {
-                socket.emit("skill_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("delete_skill", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const skillName = (data.name ?? "").trim();
-
-            if (!skillName) {
-                socket.emit("skill_result", { requestId, ok: false, message: "Missing skill name" });
-                return;
-            }
-
-            const deleted = deleteSkill(skillName);
-            const skills = scanGlobalSkills();
-            socket.emit("skill_result", {
-                requestId,
-                ok: deleted,
-                message: deleted ? undefined : "Skill not found",
-                skills,
-            });
-        });
-
-        socket.on("get_skill", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const skillName = (data.name ?? "").trim();
-            const content = skillName ? readSkillContent(skillName) : null;
-            if (content === null) {
-                socket.emit("skill_result", { requestId, ok: false, message: "Skill not found" });
-            } else {
-                socket.emit("skill_result", { requestId, ok: true, name: skillName, content });
-            }
-        });
-
-        // ── Agents management ──────────────────────────────────────────────
-
-        socket.on("list_agents", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const agents = scanGlobalAgents();
-            socket.emit("agents_list", { agents, requestId });
-        });
-
-        socket.on("create_agent", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const agentName = (data.name ?? "").trim();
-            const agentContent = data.content ?? "";
-
-            if (!agentName) {
-                socket.emit("agent_result", { requestId, ok: false, message: "Missing agent name" });
-                return;
-            }
-
-            if (!/^[a-z0-9][a-z0-9-]*[a-z0-9]$/.test(agentName) && !/^[a-z0-9]$/.test(agentName)) {
-                socket.emit("agent_result", {
-                    requestId,
-                    ok: false,
-                    message: "Invalid agent name: must be lowercase letters, numbers, and hyphens only",
-                });
-                return;
-            }
-
-            try {
-                await writeAgent(agentName, agentContent);
-                const agents = scanGlobalAgents();
-                socket.emit("agent_result", { requestId, ok: true, agents });
-            } catch (err) {
-                socket.emit("agent_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("update_agent", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const agentName = (data.name ?? "").trim();
-            const agentContent = data.content ?? "";
-
-            if (!agentName) {
-                socket.emit("agent_result", { requestId, ok: false, message: "Missing agent name" });
-                return;
-            }
-
-            // Relaxed validation for updates: accept any name the scanner
-            // can discover (letters, digits, hyphens, underscores, dots)
-            // but reject path separators to prevent directory traversal.
-            if (!/^[a-zA-Z0-9][a-zA-Z0-9._-]*$/.test(agentName)) {
-                socket.emit("agent_result", {
-                    requestId,
-                    ok: false,
-                    message: "Invalid agent name: must start with a letter or digit and contain only letters, digits, hyphens, underscores, or dots",
-                });
-                return;
-            }
-
-            try {
-                await writeAgent(agentName, agentContent);
-                const agents = scanGlobalAgents();
-                socket.emit("agent_result", { requestId, ok: true, agents });
-            } catch (err) {
-                socket.emit("agent_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("delete_agent", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const agentName = (data.name ?? "").trim();
-
-            if (!agentName) {
-                socket.emit("agent_result", { requestId, ok: false, message: "Missing agent name" });
-                return;
-            }
-
-            const deleted = deleteAgent(agentName);
-            const agents = scanGlobalAgents();
-            socket.emit("agent_result", {
-                requestId,
-                ok: deleted,
-                message: deleted ? undefined : "Agent not found",
-                agents,
-            });
-        });
-
-        socket.on("get_agent", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const agentName = (data.name ?? "").trim();
-            const content = agentName ? readAgentContent(agentName) : null;
-            if (content === null) {
-                socket.emit("agent_result", { requestId, ok: false, message: "Agent not found" });
-            } else {
-                socket.emit("agent_result", { requestId, ok: true, name: agentName, content });
-            }
-        });
-
-        // ── Plugins management ─────────────────────────────────────────────
-
-        socket.on("list_plugins", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            // Use the provided cwd (e.g. session's working directory) if
-            // available, otherwise fall back to the daemon's own cwd.
-            // Validate against workspace roots to prevent arbitrary path scanning.
-            const rawCwd = (typeof data?.cwd === "string" && data.cwd) ? data.cwd : undefined;
-            if (rawCwd && !isCwdAllowed(rawCwd)) {
-                socket.emit("plugins_list", { plugins: [], requestId, ok: false, message: "cwd outside allowed workspace roots" });
-                return;
-            }
-            // Only include project-local plugins when an explicit session cwd
-            // was provided AND it's within allowed workspace roots. Without an
-            // explicit cwd this is a runner-level query — only global plugins.
-            // When rawCwd is absent, pass undefined so marketplace discovery
-            // skips project-scoped plugins and respects only user-level settings.
-            const scanCwd = rawCwd ?? undefined;
-            const includeLocal = !!rawCwd && isCwdAllowed(rawCwd);
-            const plugins = scanAllPluginInfo(scanCwd, { includeProjectLocal: includeLocal });
-            // Echo scoped flag so the server can skip cache updates for per-session scans
-            socket.emit("plugins_list", { plugins, requestId, ...(rawCwd ? { scoped: true } : {}) });
-        });
-
-        // ── Sandbox ────────────────────────────────────────────────────────
-
-        socket.on("sandbox_get_status", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            try {
-                // Read sandbox config from disk — the daemon process does NOT
-                // run inside a sandbox itself, so process-local sandbox state
-                // (getSandboxMode, isSandboxActive, getViolations) would always
-                // return defaults.  The persisted config is the source of truth
-                // for what workers will use.
-                const { loadConfig, resolveSandboxConfig, loadGlobalConfig } = await import("../config.js");
-                // Use only the global config for resolution — the daemon
-                // process CWD may contain project-local overrides that
-                // should not influence the global settings editor.
-                const globalCfg = loadGlobalConfig();
-                const globalOnlyConfig = loadConfig(process.cwd());
-                // Override sandbox with global-only to prevent project-local
-                // settings from leaking into the status response.
-                globalOnlyConfig.sandbox = globalCfg.sandbox ?? {};
-                const resolvedConfig = resolveSandboxConfig(process.cwd(), globalOnlyConfig);
-                const mode = resolvedConfig.mode ?? "none";
-                // The daemon can't know if a worker sandbox is actively
-                // enforcing right now — report that a non-"none" mode is
-                // *configured*, not that enforcement is proven active.
-                const configured = mode !== "none";
-                socket.emit("file_result", {
-                    requestId,
-                    ok: true,
-                    mode,
-                    active: configured,
-                    configured,
-                    platform: process.platform,
-                    violations: 0,
-                    recentViolations: [],
-                    config: resolvedConfig,
-                    // Send only the *global* raw config so the UI editor
-                    // doesn't leak project-local overrides into global config
-                    // when saving.
-                    rawConfig: loadGlobalConfig().sandbox ?? {},
-                });
-            } catch (err) {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        socket.on("sandbox_update_config", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId;
-            const body = data.config;
-            try {
-                if (!body || typeof body !== "object") {
-                    socket.emit("file_result", { requestId, ok: false, message: "Invalid sandbox config body" });
-                    return;
-                }
-                const validModes = ["none", "basic", "full"];
-                if (body.mode !== undefined && !validModes.includes(body.mode)) {
-                    socket.emit("file_result", { requestId, ok: false, message: `Invalid mode "${body.mode}"` });
-                    return;
-                }
-                const { saveGlobalConfig, loadConfig, resolveSandboxConfig, loadGlobalConfig } = await import("../config.js");
-                // Merge with existing global sandbox config so UI-unmanaged
-                // fields (ignoreViolations, allowUnixSockets, proxy ports,
-                // allowGitConfig, etc.) are preserved across saves.
-                const existingSandbox = loadGlobalConfig().sandbox ?? {} as Record<string, any>;
-                // Deep-merge nested objects (filesystem, network) so that
-                // sub-fields not managed by the UI (e.g. allowGitConfig,
-                // allowUnixSockets, proxy ports) are preserved.
-                const merged: Record<string, any> = { ...existingSandbox };
-                for (const [key, value] of Object.entries(body)) {
-                    if (value && typeof value === "object" && !Array.isArray(value)
-                        && merged[key] && typeof merged[key] === "object" && !Array.isArray(merged[key])) {
-                        merged[key] = { ...merged[key], ...value };
-                    } else {
-                        merged[key] = value;
-                    }
-                }
-                saveGlobalConfig({ sandbox: merged });
-                const newConfig = loadConfig(process.cwd());
-                const resolved = resolveSandboxConfig(process.cwd(), newConfig);
-                socket.emit("file_result", {
-                    requestId,
-                    ok: true,
-                    saved: true,
-                    resolvedConfig: resolved,
-                    message: "Changes will apply on next session start.",
-                });
-            } catch (err) {
-                socket.emit("file_result", {
-                    requestId,
-                    ok: false,
-                    message: err instanceof Error ? err.message : String(err),
-                });
-            }
-        });
-
-        // ── Usage dashboard ───────────────────────────────────────────────
-
-        // ── Models ──────────────────────────────────────────────────────
-
-        socket.on("list_models", async (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data?.requestId;
-            try {
-                const models = await listConfiguredModels(process.cwd());
-                socket.emit("models_list", { requestId, models });
-            } catch (e: any) {
-                socket.emit("models_list", { requestId, models: [], error: e.message ?? "Failed to list models" });
-            }
-        });
-
-        // ── Usage ─────────────────────────────────────────────────────────
-
-        socket.on("get_usage", (data: any) => {
-            if (isShuttingDown) return;
-            const requestId = data.requestId ?? "";
-            try {
-                const range = (data.range as UsageRange) || "90d";
-                const usageData = getUsageData(range);
-                if (!usageData) {
-                    socket.emit("usage_error", {
-                        requestId,
-                        error: "Usage data not available yet — initial scan in progress",
-                    });
-                    return;
-                }
-                socket.emit("usage_data", { requestId, data: usageData });
-            } catch (e: any) {
-                socket.emit("usage_error", {
-                    requestId,
-                    error: e.message ?? "Unknown error",
-                });
-            }
-        });
+        // ── Skills / agents / plugins / sandbox / models / usage handlers ──
+        registerSkillsHandlers(socket, () => isShuttingDown);
+        registerAgentsHandlers(socket, () => isShuttingDown);
+        registerPluginsHandlers(socket, () => isShuttingDown);
+        registerSandboxHandlers(socket, () => isShuttingDown);
+        registerModelsHandlers(socket, () => isShuttingDown, listConfiguredModels);
+        registerUsageHandlers(socket, () => isShuttingDown);
 
         // ── Session Analysis ──────────────────────────────────────────
